make improvements in generation data process

+++ src/main/java/com/scoring/application/supplier/AccountSupplier.java
@@ -1,11 +1,10 @@
 package com.scoring.application.supplier;
 
-import com.scoring.domain.Account;
-import com.scoring.domain.AccountType;
-import com.scoring.domain.Client;
+import com.scoring.domain.account.Account;
+import com.scoring.domain.account.AccountType;
+import com.scoring.domain.client.Client;
 import jakarta.inject.Singleton;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.Month;
 import java.time.temporal.ChronoUnit;
@@ -16,16 +15,16 @@ import static com.scoring.application.utils.RandomUtils.*;
 @Singleton
 public class AccountSupplier {
 
-    public static final long MIN_REQUESTED_AMOUNT = 1000L;
-    public static final long MAX_REQUESTED_AMOUNT = 300000L;
+    public static final int MIN_REQUESTED_AMOUNT = 1000;
+    public static final int MAX_REQUESTED_AMOUNT = 300000;
     private static final int REQUESTED_AMOUNT_DIVIDER = 1000;
     private static final LocalDate START_INCLUSIVE = LocalDate.of(2000, Month.JANUARY, 1);
     private static final LocalDate END_EXCLUSIVE = LocalDate.of(2022, Month.DECEMBER, 31);
 
     public Account get(Client client) {
         LocalDate startDate = getRandomDate();
-        BigDecimal requestedAmount = randomBigDecimal(BigDecimal.valueOf(MIN_REQUESTED_AMOUNT), BigDecimal.valueOf(MAX_REQUESTED_AMOUNT));
-        Integer numberOfInstallments = requestedAmount.intValue() / REQUESTED_AMOUNT_DIVIDER;
+        Integer requestedAmount = randomInteger(MIN_REQUESTED_AMOUNT, MAX_REQUESTED_AMOUNT);
+        Integer numberOfInstallments = requestedAmount / REQUESTED_AMOUNT_DIVIDER;
         LocalDate endDate = getEndDate(startDate, numberOfInstallments);
 
         return Account.builder()
@@ -33,7 +32,7 @@ public class AccountSupplier {
                 .client(client)
                 .accountType(AccountType.INSTALLMENT)
                 .numberOfInstallments(numberOfInstallments)
-                .installmentAmount(BigDecimal.valueOf(REQUESTED_AMOUNT_DIVIDER))
+                .installmentAmount(REQUESTED_AMOUNT_DIVIDER)
                 .initialBalance(requestedAmount)
                 .startDate(startDate)
                 .endDate(endDate)
